+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - `collapsible-list`
   - [#4623](https://github.com/wix/ricos/pull/4623) render plugin menu inside collapsible list body
+- `toolbars-v3`
+  - [#4631](https://github.com/wix/ricos/pull/4631) support sideToolbar onClick config
 
 </details>